requirements.txt
streamlit==1.25.0
opencv-python-headless==4.8.0.76  # OpenCV without GUI functionality (for efficient deployment)
numpy==1.23.5
tensorflow==2.13.0
Pillow==9.2.0
